% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_expectations.R
\name{make_expectations}
\alias{make_expectations}
\title{Make an expectation list from rmarkdown file}
\usage{
make_expectations(file, keyword_prefix = "### ")
}
\arguments{
\item{file}{an rmarkdown file or an r charactor vector.}

\item{keyword_prefix}{a common part of all keywords. By default is "### ".}
}
\description{
This function takes a file, searches all kye_words using key_word_prefix argument and then create a list with evaluation of all chunks with keywords.
}
\examples{
make_expectations(test_1)
}
\author{
George Moroz <agricolamz@gmail.com>
}

