% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistaken_tasks.R
\name{mistaken_tasks}
\alias{mistaken_tasks}
\title{Creates rmarkdown from mestaken tasks}
\usage{
mistaken_tasks(results, keyword_prefix = "### ", write_file = TRUE,
  show_file = TRUE)
}
\arguments{
\item{results}{data frame with test_files() results}

\item{keyword_prefix}{a common part of all keywords. By default is "### ".}

\item{write_file}{logical. If TRUE, writes the created rmarkdown to the working directory.}

\item{show_file}{logical. If TRUE, opens in RStudio the created rmarkdown.}
}
\description{
This function takes a results of test_files() function and writes in working directory an rmarkdown containing all mistaken tasks made by students.
}
\examples{
my_files <- c(
system.file("extdata", "test_1.Rmd", package = "AssignmentCheck"),
system.file("extdata", "test_2.Rmd", package = "AssignmentCheck"),
system.file("extdata", "test_3.Rmd", package = "AssignmentCheck"),
system.file("extdata", "test_4.Rmd", package = "AssignmentCheck")
)
fit <- lm(mpg~cyl, mtcars)
result <- test_files(expectations = list(2+2, fit$coefficients), files = my_files)
mistaken_tasks(results = result, write_file = FALSE)
}
\author{
George Moroz <agricolamz@gmail.com>
}

