% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_task.R
\name{test_task}
\alias{test_task}
\title{Test rmarkdown task}
\usage{
test_task(expectations, file, keyword)
}
\arguments{
\item{expectations}{an expected value.}

\item{file}{an rmarkdown file or an r charactor vector.}

\item{keyword}{a key_word to identify, where shuld be r chunk searched.}
}
\description{
This function takes a file, searches a kye_word line and then evaluate the first r chunk after the keyword. Then function compare ther result with the expected value.
}
\examples{
test_task(expectations = 2+2, file = test_1, keyword = "### 1.1")
test_task(expectations = 2+3, file = test_1, keyword = "### 1.1")
}
\author{
George Moroz <agricolamz@gmail.com>
}

