% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_files.R
\name{test_files}
\alias{test_files}
\title{Test all tasks in all rmarkdown files
This function takes all files in directory, searches all kye_words using key_word_prefix argument and then tests associated with all keywords r chunks. Then function compare the results with the expected value and return a dataframe containing file name, author, check results and number of corrected answers.}
\usage{
test_files(expectations, files = list.files(), keyword_prefix = "### ",
  group = FALSE)
}
\arguments{
\item{expectations}{an expected value.}

\item{files}{an rmarkdown files. By dafault is all .Rmd files in a working directory.}

\item{keyword_prefix}{a common part of all keywords. By default is "### ".}

\item{group}{logical. If TRUE, function searches first line with the "group:" argument and returns it in the summary.}
}
\description{
Test all tasks in all rmarkdown files
This function takes all files in directory, searches all kye_words using key_word_prefix argument and then tests associated with all keywords r chunks. Then function compare the results with the expected value and return a dataframe containing file name, author, check results and number of corrected answers.
}
\examples{
\dontrun{
fit <- lm(mpg~cyl, mtcars)
test_files(expectations = list(2+2, fit$coefficients))}
}
\author{
George Moroz <agricolamz@gmail.com>
}

